[UI] Android NDK portability + Checked -Wformat-security fixes
ImGui printf-family non-literal calls -> TextUnformatted/%s (xam_ui, updater_dialog, netplay_settings_dialog, profile_dialogs, gamerpic_browser); game_detail_panel %llu cast on LP64; gamerpic_browser AtomicSharedPtr for atomic<shared_ptr> NDK gap; updater execlp guarded to !XE_PLATFORM_ANDROID; discord_presence no-op on Android.

+++ src/xenia/app/game_detail_panel.cc
@@ -61,7 +61,8 @@ void GameDetailPanel::Draw(float width) {
     ImGui::TextDisabled("Type: %s", entry_->media_type.c_str());
   }
   if (entry_->play_seconds > 0) {
-    ImGui::Text("Play time: %llu min", entry_->play_seconds / 60);
+    ImGui::Text("Play time: %llu min",
+                static_cast<unsigned long long>(entry_->play_seconds / 60));
   }
 
   if (auto meta = metadata_.Lookup(entry_->title_id)) {
